Split the logic to parse comments out of Comment.php
Leave Comment::fromStringInContext() and methods to access comments
inside of the class Context.
